sw: Rename barriers and move global barrier into .dram section
