+++ sw/snRuntime/src/start_banshee.c
@@ -31,12 +31,6 @@ void _snrt_init_team(uint32_t cluster_core_id, uint32_t cluster_core_num,
     team->cluster_mem.start = spm_start;
     team->cluster_mem.end = spm_end;
 
-    // Allocate and initialize global barrier
-    team->global_barrier = (void *)team->global_mem.start;
-    team->global_barrier->barrier = 0;
-    team->global_barrier->barrier_iteration = 0;
-    team->global_mem.start += sizeof(struct snrt_barrier);
-
     // Allocate memory for a global mailbox.
     team->global_mailbox = team->global_mem.start;
     team->global_mem.start += sizeof(struct snrt_mailbox);
