+++ sw/snRuntime/src/start_cluster.c
@@ -28,7 +28,7 @@ struct snrt_cluster_bootdata {
 };
 
 // Rudimentary string buffer for putc calls.
-extern uint32_t _end;
+extern uint32_t _edram;
 #define PUTC_BUFFER_LEN (1024 - sizeof(size_t))
 struct putc_buffer_header {
     size_t size;
@@ -37,7 +37,7 @@ struct putc_buffer_header {
 static volatile struct putc_buffer {
     struct putc_buffer_header hdr;
     char data[PUTC_BUFFER_LEN];
-} *const putc_buffer = (void *)&_end;
+} *const putc_buffer = (void *)&_edram;
 
 void _snrt_init_team(uint32_t cluster_core_id, uint32_t cluster_core_num,
                      void *spm_start, void *spm_end,
@@ -63,12 +63,6 @@ void _snrt_init_team(uint32_t cluster_core_id, uint32_t cluster_core_num,
     // Initialize cluster barrier
     team->cluster_barrier.barrier = 0;
     team->cluster_barrier.barrier_iteration = 0;
-
-    // Allocate and initialize global barrier
-    team->global_barrier = (void *)team->global_mem.start;
-    team->global_barrier->barrier = 0;
-    team->global_barrier->barrier_iteration = 0;
-    team->global_mem.start += sizeof(struct snrt_barrier);
 
     // Allocate memory for a global mailbox.
     team->global_mailbox = (void *)team->global_mem.start;
