+++ sw/snRuntime/src/team.c
@@ -99,57 +99,16 @@ void snrt_bcast_send(void *data, size_t len) {
     struct snrt_mailbox *mbox = get_mailbox();
     mbox->ptr = data;
     mbox->len = len;
-    snrt_barrier();
+    snrt_cluster_hw_barrier();
     // Others read here.
-    snrt_barrier();
+    snrt_cluster_hw_barrier();
 }
 
 /// Receive a chunk of data broadcast from another core in the team.
 void snrt_bcast_recv(void *data, size_t len) {
     struct snrt_mailbox *mbox = get_mailbox();
     // Main core writes here.
-    snrt_barrier();
+    snrt_cluster_hw_barrier();
     snrt_memcpy(data, mbox->ptr, len);
-    snrt_barrier();
-}
-
-/// Synchronize cores in a cluster with a software barrier
-void snrt_cluster_barrier() {
-    // Remember previous iteration
-    struct snrt_barrier *barrier_ptr =
-        &_snrt_team_current->root->cluster_barrier;
-    uint32_t prev_barrier_iteration = barrier_ptr->barrier_iteration;
-    uint32_t barrier =
-        __atomic_add_fetch(&barrier_ptr->barrier, 1, __ATOMIC_RELAXED);
-
-    // Increment the barrier counter
-    if (barrier == snrt_cluster_core_num()) {
-        barrier_ptr->barrier = 0;
-        __atomic_add_fetch(&barrier_ptr->barrier_iteration, 1,
-                           __ATOMIC_RELAXED);
-    } else {
-        // Some threads have not reached the barrier --> Let's wait
-        while (prev_barrier_iteration == barrier_ptr->barrier_iteration)
-            ;
-    }
-}
-
-/// Synchronize clusters globally with a global software barrier
-void snrt_global_barrier() {
-    // Remember previous iteration
-    struct snrt_barrier *barrier_ptr = _snrt_team_current->root->global_barrier;
-    uint32_t prev_barrier_iteration = barrier_ptr->barrier_iteration;
-    uint32_t barrier =
-        __atomic_add_fetch(&barrier_ptr->barrier, 1, __ATOMIC_RELAXED);
-
-    // Increment the barrier counter
-    if (barrier == snrt_global_core_num()) {
-        barrier_ptr->barrier = 0;
-        __atomic_add_fetch(&barrier_ptr->barrier_iteration, 1,
-                           __ATOMIC_RELAXED);
-    } else {
-        // Some threads have not reached the barrier --> Let's wait
-        while (prev_barrier_iteration == barrier_ptr->barrier_iteration)
-            ;
-    }
+    snrt_cluster_hw_barrier();
 }
