+++ sw/snRuntime/src/team.h
@@ -50,6 +50,5 @@ struct snrt_team_root {
     struct snrt_mailbox *cluster_mailbox;
     struct snrt_allocator allocator;
     struct snrt_barrier cluster_barrier;
-    struct snrt_barrier *global_barrier;
     uint32_t barrier_reg_ptr;
 };
